chore: keep ElasticInferencePool in elastic.py for smaller diff
